# Single precision support

Fixes #1079, #1083

---
# Change Description

Added a build-time option to switch to `float` (single precision) for
all computations. The generic type `FloatType`, previously used
inconsistently and in the dynamics only, was introduced everywhere to
replace `double`. To make it compile, some double literals had to be
wrapped (e.g. `FloatType(0.5)`). There are likely more places where
double literals cause unnecessary promotion of an operation to `double`.

Some bugs, unrelated to the float type, where fixed while debugging
single precision:

* Kokkos with a host backend was broken because of incorrect buffer
lifetimes
* `ssh` field was overwritten in the Kokkos kernel with 0 before
computations
* `SlabOcean` computation was incorrect due to parentheses placement 
* `SlabOcean` test initialization was missing some fields since the
introduction of Accessors
* kokkos host backend could not handle reallocation the underlying
ModelArray buffer which happens in some tests

---
# Test Description
The regular tests all pass when built in single precision. To achieve
this, some error tolerances had to be adjusted based on the floating
point type because of the reduced accuracy of `float`. More substantial
changes where needed in the `slabOcean`, where the reference
computations had to be updated to match the fixed equation in the
implementation. @timspainNERSC Can you have a look again at these
changes in the test?

Both the benchmark and the NH case where run for a few simulated days in
single precision and look fine.

---
# Documentation Impact
This feature is mentioned in the Installation/GPU build section. 

---
# Other Details

+++ core/src/FieldAdvection.cpp
@@ -13,7 +13,7 @@
 namespace Nextsim {
 
 ModelArray& FieldAdvection::advectField(
-    ModelArray& field, const TimestepTime& tst, double lowerLimit, double upperLimit)
+    ModelArray& field, const TimestepTime& tst, FloatType lowerLimit, FloatType upperLimit)
 {
     Module::getImplementation<IDynamics>().advectField(
         tst.step.seconds(), field, lowerLimit, upperLimit);
@@ -22,7 +22,7 @@ ModelArray& FieldAdvection::advectField(
 
 #ifdef USE_KOKKOS
 void FieldAdvection::advectField(
-    const DeviceViewMA& field, const TimestepTime& tst, double lowerLimit, double upperLimit)
+    const DeviceViewMA& field, const TimestepTime& tst, FloatType lowerLimit, FloatType upperLimit)
 {
     Module::getImplementation<IDynamics>().advectField(
         tst.step.seconds(), field, lowerLimit, upperLimit);
