+++ physics/src/modules/AtmosphereBoundaryModule/include/ConfiguredAtmosphere.hpp
@@ -41,15 +41,15 @@ public:
     void update(const TimestepTime&) override;
 
 private:
-    static double tair0;
-    static double tdew0;
-    static double pair0;
-    static double sw0;
-    static double lw0;
-    static double snowfall0;
-    static double rain0;
-    static double uWind0;
-    static double vWind0;
+    static FloatType tair0;
+    static FloatType tdew0;
+    static FloatType pair0;
+    static FloatType sw0;
+    static FloatType lw0;
+    static FloatType snowfall0;
+    static FloatType rain0;
+    static FloatType uWind0;
+    static FloatType vWind0;
 
     ModelArrayAccessor<Protected::T_AIR, RW> tairAccessor;
     ModelArrayAccessor<Protected::DEW_2M, RW> tdewAccessor;
