+++ physics/src/modules/IceAlbedoModule/include/SMUIceAlbedo.hpp
@@ -22,7 +22,7 @@ public:
     void update(const TimestepTime& tst) override;
 
 private:
-    static double i0;
+    static FloatType i0;
 };
 
 }
